Refactor mail_draft tests into two files
