+++ handoverservice/mail_draft/tests_mailer.py
@@ -1,7 +1,6 @@
 from django.test import TestCase
 from mail_draft.mailer import generate_message
 from handover_api.models import User
-import mock
 
 class MailerTestCase(TestCase):
 
@@ -28,21 +27,3 @@ class MailerTestCase(TestCase):
         self.assertEqual(sender_email, message.from_email)
         self.assertIn(rcpt_email, message.to)
 
-
-class DDSUtilTestCase(TestCase):
-    @mock.patch('ddsc.core.remotestore.RemoteStore')
-    def testGetEmail(self, mockRemoteStore):
-        user_id = 'abcd-1234-efgh-8876'
-        email = 'example@domain.com'
-        # Mock a remote user object, and bind it to fetch_user
-        remote_user = mock.Mock()
-        remote_user.email = email
-        instance = mockRemoteStore.return_value
-        instance.fetch_user.return_value = remote_user
-        # Only import DDSUtil once we've patched RemoteStore
-        from dds_util import DDSUtil
-        User.objects.create(dds_id=user_id, api_key='uhn3wk7h24ighg8i2')
-        # DDSUtil reads settings from django settings, so inject some here
-        with self.settings(DDSCLIENT_PROPERTIES={}):
-            dds_util = DDSUtil(user_id)
-            self.assertEqual(email, dds_util.get_email_address(user_id))
